api/requirements.txt
flask
pytest
psycopg2-binary
simplejson
intrinio_sdk
finnhub-python
